requirements.txt
boto==2.49.0
boto3==1.14.41
Django==2.2
djangorestframework==3.9.0
django-cors-headers==2.4.0
django-constance==2.2.0
django-debug-toolbar==2.2
django-dynamic-decimal==0.3.0
django-import-export==2.5
django-fluent-contents==2.0.7
django-extensions==2.2.1
django-filer==1.7.1
django-modeltranslation==0.15
django-picklefield==1.0.0
django-sekizai==0.10.0
django-smart-selects==1.5.3
django-redis-cache==2.1.1
django-tenants==3.1.0
django-storages==1.10.1
django-treebeard==4.2.0
django-tinymce==2.7.0
django-webpack-loader==0.6.0
psycopg2==2.7.5
easy-thumbnails==2.6
pylibmc==1.6.1
python-memcached==1.59
pyaml==15.8.2
PyYAML==5.3.1
gunicorn==19.6.0
redis==3.5.3
sqlparse==0.2.4
Unidecode==0.4.21
pytz==2018.9
django-polymorphic==2.1.2
django-cors-headers==2.4.0
django-environ==0.4.5
django-parler==2.2
django-parler-rest==2.1
Werkzeug==0.16.0
dj-database-url==0.5.0
django-maintenance-mode==0.16.1
django-allow-cidr==0.3.1
requests==2.28.2